Move blink service registration
