+++ homeassistant/components/blink/camera.py
@@ -9,35 +9,23 @@ from typing import Any
 from blinkpy.auth import UnauthorizedError
 from blinkpy.camera import BlinkCamera as BlinkCameraAPI
 from requests.exceptions import ChunkedEncodingError
-import voluptuous as vol
 
 from homeassistant.components.camera import Camera
-from homeassistant.const import CONF_FILE_PATH, CONF_FILENAME
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import (
     ConfigEntryAuthFailed,
     HomeAssistantError,
     ServiceValidationError,
 )
-from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import (
-    DEFAULT_BRAND,
-    DOMAIN,
-    SERVICE_RECORD,
-    SERVICE_SAVE_RECENT_CLIPS,
-    SERVICE_SAVE_VIDEO,
-    SERVICE_TRIGGER,
-)
+from .const import DEFAULT_BRAND, DOMAIN
 from .coordinator import BlinkConfigEntry, BlinkUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
-ATTR_VIDEO_CLIP = "video"
-ATTR_IMAGE = "image"
 PARALLEL_UPDATES = 1
 
 
@@ -55,20 +43,6 @@ async def async_setup_entry(
     ]
 
     async_add_entities(entities)
-
-    platform = entity_platform.async_get_current_platform()
-    platform.async_register_entity_service(SERVICE_RECORD, None, "record")
-    platform.async_register_entity_service(SERVICE_TRIGGER, None, "trigger_camera")
-    platform.async_register_entity_service(
-        SERVICE_SAVE_RECENT_CLIPS,
-        {vol.Required(CONF_FILE_PATH): cv.string},
-        "save_recent_clips",
-    )
-    platform.async_register_entity_service(
-        SERVICE_SAVE_VIDEO,
-        {vol.Required(CONF_FILENAME): cv.string},
-        "save_video",
-    )
 
 
 class BlinkCamera(CoordinatorEntity[BlinkUpdateCoordinator], Camera):
