+++ homeassistant/components/blink/const.py
@@ -20,11 +20,6 @@ TYPE_TEMPERATURE = "temperature"
 TYPE_BATTERY = "battery"
 TYPE_WIFI_STRENGTH = "wifi_strength"
 
-SERVICE_RECORD = "record"
-SERVICE_TRIGGER = "trigger_camera"
-SERVICE_SAVE_VIDEO = "save_video"
-SERVICE_SAVE_RECENT_CLIPS = "save_recent_clips"
-SERVICE_SEND_PIN = "send_pin"
 
 PLATFORMS = [
     Platform.ALARM_CONTROL_PANEL,
