+++ homeassistant/components/blink/services.py
@@ -4,13 +4,35 @@ from __future__ import annotations
 
 import voluptuous as vol
 
-from homeassistant.const import ATTR_CONFIG_ENTRY_ID, CONF_PIN
+from homeassistant.components.camera import DOMAIN as CAMERA_DOMAIN
+from homeassistant.const import (
+    ATTR_CONFIG_ENTRY_ID,
+    CONF_FILE_PATH,
+    CONF_FILENAME,
+    CONF_NAME,
+    CONF_PIN,
+)
 from homeassistant.core import HomeAssistant, ServiceCall, callback
 from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers import config_validation as cv, issue_registry as ir
+from homeassistant.helpers import config_validation as cv, issue_registry as ir, service
 
-from .const import DOMAIN, SERVICE_SEND_PIN
+from .const import DOMAIN
 
+SERVICE_RECORD = "record"
+SERVICE_TRIGGER = "trigger_camera"
+SERVICE_SAVE_VIDEO = "save_video"
+SERVICE_SAVE_RECENT_CLIPS = "save_recent_clips"
+SERVICE_SEND_PIN = "send_pin"
+
+SERVICE_SAVE_VIDEO_SCHEMA = vol.Schema(
+    {vol.Required(CONF_NAME): cv.string, vol.Required(CONF_FILENAME): cv.string}
+)
+SERVICE_SAVE_RECENT_CLIPS_SCHEMA = vol.Schema(
+    {vol.Required(CONF_NAME): cv.string, vol.Required(CONF_FILE_PATH): cv.string}
+)
+
+# Deprecated
+SERVICE_SEND_PIN_SCHEMA = vol.Schema({vol.Optional(CONF_PIN): cv.string})
 SERVICE_SEND_PIN_SCHEMA = vol.Schema(
     {
         vol.Required(ATTR_CONFIG_ENTRY_ID): vol.All(cv.ensure_list, [cv.string]),
@@ -51,4 +73,37 @@ def async_setup_services(hass: HomeAssistant) -> None:
         SERVICE_SEND_PIN,
         _send_pin,
         schema=SERVICE_SEND_PIN_SCHEMA,
+    )
+
+    service.async_register_platform_entity_service(
+        hass,
+        DOMAIN,
+        SERVICE_RECORD,
+        entity_domain=CAMERA_DOMAIN,
+        schema=None,
+        func="record",
+    )
+    service.async_register_platform_entity_service(
+        hass,
+        DOMAIN,
+        SERVICE_TRIGGER,
+        entity_domain=CAMERA_DOMAIN,
+        schema=None,
+        func="trigger_camera",
+    )
+    service.async_register_platform_entity_service(
+        hass,
+        DOMAIN,
+        SERVICE_SAVE_RECENT_CLIPS,
+        entity_domain=CAMERA_DOMAIN,
+        schema={vol.Required(CONF_FILE_PATH): cv.string},
+        func="save_recent_clips",
+    )
+    service.async_register_platform_entity_service(
+        hass,
+        DOMAIN,
+        SERVICE_SAVE_VIDEO,
+        entity_domain=CAMERA_DOMAIN,
+        schema={vol.Required(CONF_FILENAME): cv.string},
+        func="save_video",
     )
